+++ sphinx/source/changelog.rst
@@ -5,7 +5,19 @@ Changelog (Ren'Py 7.x-)
 *There is also a list of* :doc:`incompatible changes <incompatible>`
 
 
+.. _renpy-8.5.4:
+
+8.5.4
+=====
+
+
+The interpretation of Bezier curves in Live2D has been changed to match the Cardano interpretation,
+which is consistent with the Live2D SDK. This behavior can be changed with the `old_beziers` parameter to the :func:`Live2D`
+displayable, or with the :var:`config.live2d_old_beziers` variable,
+
+
 .. _renpy-8.5.3:
+
 
 8.5.3
 =====